refactor(payment): 결제 가격을 Integer에서 Long으로 변경

+++ src/main/java/staysplit/hotel_reservation/reservation/domain/entity/ReservationParticipantEntity.java
@@ -27,7 +27,7 @@ public class ReservationParticipantEntity {
     private CustomerEntity customer;
 
     @Column(nullable = false)
-    private Integer splitAmount;
+    private Long splitAmount;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 20)
